# web/public/openapi.yaml
openapi: 3.0.3
info:
  title: Pemuda Surau Al-Abqori API
  version: 1.0.0
  description: |
    REST API for the Surau Al-Abqori youth programme: prayer attendance
    (face-verified + geofenced), Quran activity, merits, programs and the
    public dashboard.

    ## Authentication
    Most endpoints require a JWT obtained from `POST /api/auth/verify-otp`.
    Send it as `Authorization: Bearer <token>`.

    ## Authorisation
    Beyond "logged in", endpoints are gated by **capabilities** rather than raw
    roles (see `server/src/utils/roles.js`). Admins implicitly hold every
    capability. The `x-capability` extension on each operation records the
    requirement.

    ## Geofencing
    Check-in, activity recording and new registrations are only accepted within
    a radius of the surau. Coordinates are supplied by the client.
servers:
  - url: /api
    description: Same origin (proxied by Nginx / the Vite dev server)
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Authentication required
        code:
          type: string
          description: Optional machine-readable code (e.g. REGISTRATION_TOO_FAR).
          example: REGISTRATION_TOO_FAR
        retryAfterMinutes:
          type: integer
          description: Present on 429 responses (e.g. the Quran log cooldown).
...
    Descriptor:
      type: array
      description: 128-float face descriptor produced by face-api.js in the browser.
      items: { type: number }
      minItems: 128
      maxItems: 128
    Attendance:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        prayer: { type: string, enum: [subuh, zuhur, asar, maghrib, isyak] }
        attendance_date: { type: string, format: date }
        checked_in_at: { type: string, format: date-time }
        method: { type: string, enum: [face, manual] }
        face_score: { type: number, nullable: true }
        latitude: { type: number, nullable: true }
        longitude: { type: number, nullable: true }
    PrayerWindow:
      type: object
      nullable: true
      properties:
        prayer: { type: string, enum: [subuh, zuhur, asar, maghrib, isyak] }
        adhan: { type: string, format: date-time }
        start: { type: string, format: date-time }
        end: { type: string, format: date-time }
...
  responses:
...
    Forbidden:
      description: Authenticated but lacking the required capability.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
paths:
...
  /attendance/current:
    get:
      tags: [Attendance]
      summary: Which prayer is open for check-in
      description: |
        Returns the currently open window (if any) and the next one, so the
        client can show a countdown. Returns 503 if prayer times cannot be
        fetched from the upstream provider.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  date: { type: string, format: date }
                  enforceWindow: { type: boolean }
                  windowBeforeMinutes: { type: integer }
                  windowAfterMinutes: { type: integer }
                  current: { $ref: '#/components/schemas/PrayerWindow' }
                  next: { $ref: '#/components/schemas/PrayerWindow' }
        '503':
          description: Prayer times temporarily unavailable.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /attendance/check-in:
    post:
      tags: [Attendance]
      summary: Face-verified prayer check-in
      description: |
        The prayer is inferred server-side from the current window. The device
        must be within the check-in geofence (default 150 m). A guardian may
        check in a dependent by passing `forUserId`; the scanned face must then
        match that dependent.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [descriptor, latitude, longitude]
              properties:
                descriptor: { $ref: '#/components/schemas/Descriptor' }
                latitude: { type: number }
                longitude: { type: number }
                forUserId:
                  type: string
                  format: uuid
                  description: A dependent of the caller. Omit to check in yourself.
                prayer:
                  type: string
                  description: Only honoured when window enforcement is disabled.
      responses:
        '200':
          description: Checked in
          content:
            application/json:
              schema:
                type: object
                properties:
                  attendance: { $ref: '#/components/schemas/Attendance' }
                  confidence: { type: number }
        '401':
          description: Face not recognized.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: Outside the geofence, or the face does not match the selected member.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: No prayer window is open, or the member has no face enrolled.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /attendance/identify:
    post:
      tags: [Attendance]
      summary: Identify a member by face (staff)
      description: |
        Requires the `identifyMembers` capability. Returns a *suggestion* only —
        the client must confirm before anything is recorded.
      x-capability: identifyMembers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [descriptor, latitude, longitude]
              properties:
                descriptor: { $ref: '#/components/schemas/Descriptor' }
                latitude: { type: number }
                longitude: { type: number }
      responses:
        '200':
          description: Best match (and runner-up)
          content:
            application/json:
              schema:
                type: object
                properties:
                  match:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
                      fullName: { type: string }
                      isDependent: { type: boolean }
                      confidence: { type: number }
                  runnerUp:
                    type: object
                    nullable: true
                    properties:
                      fullName: { type: string }
                      confidence: { type: number }
                  distanceFromSurau: { type: integer, description: Metres. }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /attendance/me:
    get:
      tags: [Attendance]
      summary: My attendance history
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 60 }
      responses:
        '200': { description: OK }
...
  /attendance/family:
    get:
      tags: [Attendance]
      summary: Attendance history for me and my dependents
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 60 }
      responses:
        '200': { description: OK }
  /attendance/manual:
    post:
      tags: [Attendance]
      summary: Manual check-in (admin)
      description: Admin only. Records attendance on a member's behalf, e.g. when face verification failed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, prayer, date]
              properties:
                userId: { type: string, format: uuid }
                prayer: { type: string, enum: [subuh, zuhur, asar, maghrib, isyak] }
                date: { type: string, format: date }
      responses:
        '200': { description: Recorded }
  /attendance/date/{date}:
    get:
      tags: [Attendance]
      summary: Attendance for a date (admin)
      parameters:
        - name: date
          in: path
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  attendance:
                    type: array
                    items: { $ref: '#/components/schemas/Attendance' }

  # -------------------------------------------------------------- Activity
...